Backend/requirements.txt
flask
pandas
yfinance
scikit-learn
joblib
shap
requests
plotly
altair
python-dotenv
uvicorn
fastapi